monotonic align

+++ optispeech/dataset/feature_extractors/__init__.py
@@ -194,7 +194,7 @@ class CommonFeatureExtractor(FeatureExtractor):
             )
         )
 
-        magnitudes = torch.sqrt(spec.pow(2).sum(-1) + (1e-9))
+        magnitudes = torch.sqrt(spec.pow(2).sum(-1) + (1e-9)).float()
         spec = torch.matmul(self.mel_basis[mel_basis_key], magnitudes)
         spec = spectral_normalize_torch(spec)
         return spec.squeeze().cpu().numpy()
